+++ utils/lite-components-testkit/src/main/scala/pl/touk/nussknacker/engine/lite/util/test/SynchronousLiteInterpreter.scala
@@ -15,8 +15,8 @@ import pl.touk.nussknacker.engine.lite.api.customComponentTypes.CapabilityTransf
 import pl.touk.nussknacker.engine.lite.api.interpreterTypes.{EndResult, ScenarioInputBatch}
 import pl.touk.nussknacker.engine.lite.api.runtimecontext.LiteEngineRuntimeContextPreparer
 import pl.touk.nussknacker.engine.lite.capabilities.FixedCapabilityTransformer
-import pl.touk.nussknacker.engine.resultcollector.ProductionServiceInvocationCollector
 import pl.touk.nussknacker.engine.util.SynchronousExecutionContext
+import pl.touk.nussknacker.engine.util.test.TestScenarioCollectorHandler
 
 import scala.concurrent.duration._
 import scala.concurrent.{Await, ExecutionContext, Future}
@@ -49,13 +49,13 @@ object SynchronousLiteInterpreter {
       componentUseCase: ComponentUseCase,
       runtimeContextPreparer: LiteEngineRuntimeContextPreparer = LiteEngineRuntimeContextPreparer.noOp
   ): SynchronousResult = {
-
+    val testScenarioCollectorHandler = TestScenarioCollectorHandler.createHandler(componentUseCase)
     ScenarioInterpreterFactory
       .createInterpreter[Id, Any, AnyRef](
         scenario,
         modelData,
         Nil,
-        ProductionServiceInvocationCollector,
+        testScenarioCollectorHandler.resultCollector,
         componentUseCase
       )
       .map { interpreter =>
@@ -64,6 +64,7 @@ object SynchronousLiteInterpreter {
           val value: Id[ResultType[EndResult[AnyRef]]] = interpreter.invoke(data)
           value.run
         } finally {
+          testScenarioCollectorHandler.close()
           interpreter.close()
         }
       }